move page into workspace folder
